Pattern Command

+++ src/main/java/service/UserServiceImpl.java
@@ -1,34 +1,53 @@
 package service;
 
-import dao.UserDao;
+import dao.UserService;
 import entity.User;
 
 import java.sql.SQLException;
 import java.util.List;
 
-public class UserServiceImpl implements UserService {
-    private final UserDao userDao;
+public class UserServiceImpl implements service.UserService {
+    private final UserService userDao;
 
-    public UserServiceImpl(UserDao userDao) {
+    public UserServiceImpl(UserService userDao) {
         this.userDao = userDao;
     }
 
     @Override
     public void insertUser(User user) {
-        User newUser = User.builder()
-                .userName(user.getUserName())
-                .timeToSpend(user.getTimeToSpend())
-                .activity(user.getActivity())
-                .timeToAdd(user.getTimeToAdd())
-                .build();
-        userDao.insertUser(newUser);
+        try{
+            if (user == null) throw new IllegalArgumentException();
+            if (user.getTimeToSpend() < 0 || user.getTimeToSpend() > 24) throw new NumberFormatException();
+
+
+                User newUser = User.builder()
+                        .userName(user.getUserName())
+                        .timeToSpend(user.getTimeToSpend())
+                        .activity(user.getActivity())
+                        .timeToAdd(user.getTimeToAdd())
+                        .build();
+                userDao.insertUser(newUser);
+
+        }catch (NumberFormatException exception){
+            throw new NumberFormatException("incorrect time");
+        }
+        catch (IllegalArgumentException e){
+            throw new IllegalArgumentException("not found user");
+        }
     }
 
     @Override
     public User selectUser(int id) {
-        return userDao.selectUser(id);
-    }
+        User user = userDao.selectUser(id);
+        try {
+            if (user == null) throw new NullPointerException();
 
+            return userDao.selectUser(id);
+
+        }catch (NullPointerException e){
+            throw new NullPointerException("not found user");
+        }
+    }
     @Override
     public List<User> selectAllUsers() {
         return userDao.selectAllUsers();
@@ -36,11 +55,32 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public boolean deleteUser(int id) throws SQLException {
-        return userDao.deleteUser(id);
+        User user = userDao.selectUser(id);
+        try {
+            if (user == null) throw new NullPointerException();
+
+            return userDao.deleteUser(id);
+
+        }catch (NullPointerException e){
+            throw new NullPointerException("not found user");
+        }
     }
 
     @Override
     public boolean updateUser(User user) throws SQLException {
-        return userDao.updateUser(user);
+        User newUser = userDao.selectUser(user.getId());
+        try {
+            if (newUser == null) throw new IllegalArgumentException();
+            if (user.getTimeToSpend() < 0 || user.getTimeToSpend() > 24) throw new NumberFormatException();
+
+            return userDao.updateUser(user);
+
+        }catch (NumberFormatException exception){
+            throw new NumberFormatException("incorrect time");
+        }
+        catch (IllegalArgumentException e){
+            throw new IllegalArgumentException("not found user");
+        }
+
     }
 }
